requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
PyJWT==2.10.1
httpx==0.27.2
python-dateutil==2.9.0.post0
